+++ src/constants/default-graphql-sdl-file-header.ts
@@ -1,0 +1,9 @@
+const DEFAULT_GRAPHQL_SDL_FILE_HEADER = `
+#
+# -------------------------------------------------------
+# THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
+# -------------------------------------------------------
+#
+`.trim();
+
+export default DEFAULT_GRAPHQL_SDL_FILE_HEADER;