+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.16](https://github.com/abdulghani/graphql-generator/compare/v0.1.16-rc.0...v0.1.16) (2022-02-13)
+
+### 0.1.16-rc.0 (2022-02-13)
+
+
+### Features
+
+* add  to generate types to use graphql resolver argument objects using custom field resolver ([6df0cb1](https://github.com/abdulghani/graphql-generator/commit/6df0cb1e12b4b0b3e8d1b6626c77c2def248e66c))
+
+### [0.1.15](https://github.com/abdulghani/graphql-generator/compare/v0.1.15-rc.4...v0.1.15) (2022-02-13)
+
+### [0.1.15-rc.4](https://github.com/abdulghani/graphql-generator/compare/v0.1.13...v0.1.15-rc.4) (2022-02-13)
+
+
+### Features
+
+* add graphql context type path ([60e6b33](https://github.com/abdulghani/graphql-generator/commit/60e6b33b447705ad14b69db660b15acb20bc00ae))
+
+### 0.1.15-rc.3 (2022-02-13)
+
+### 0.1.15-rc.2 (2022-02-13)
+
+### 0.1.15-rc.1 (2022-02-13)
+
+### 0.1.15-rc.0 (2022-02-13)
+
+
+### Features
+
+* add graphql context type path ([60e6b33](https://github.com/abdulghani/graphql-generator/commit/60e6b33b447705ad14b69db660b15acb20bc00ae))
+
 ### [0.1.14](https://github.com/abdulghani/graphql-generator/compare/v0.1.14-rc.0...v0.1.14) (2022-02-12)
 
 ### 0.1.14-rc.0 (2022-02-12)